Do not `set :hidden`, it's really annoying
Set hidden won't delete .swp file even the buffer is unloaded.
This will be be very annoying. For example:

You open `Gemfile` in a tmux window, modify and save it and switch to another
file. Then if you try to edit Gemfile in another window, it will
warn you the swap file exists and you have to `edit anyway` or `read only`